+++ Axiom-Editor/src/Gui/EditorIcons.hpp
@@ -2,6 +2,7 @@
 #include "Graphics/Texture2D.hpp"
 #include "Serialization/Path.hpp"
 #include "Core/Export.hpp"
+#include <cstdint>
 #include <string>
 #include <unordered_map>
 
@@ -11,10 +12,12 @@ namespace Axiom {
 		static void Initialize();
 		static void Shutdown();
 
-		/// Returns the OpenGL texture handle for the given icon at the best size.
+		/// Returns the backend texture handle for the given icon at the best size.
+		/// Under WebGPU this is the raw WGPUTextureView pointer (cast to uint64_t)
+		/// so callers can pass it straight to ImGui::Image as an ImTextureID.
 		/// @param name  Icon name, e.g. "play", "stop", "open_folder"
 		/// @param size  Desired pixel size (snaps to nearest available: 16,24,32,48,64,128,256,512)
-		static unsigned int Get(const std::string& name, int size = 32);
+		static uint64_t Get(const std::string& name, int size = 32);
 
 	private:
 		struct IconEntry {
